Partially extract module setup into a script plugin
