+++ build.gradle.kts
@@ -26,27 +26,9 @@
 
 @file:Suppress("RemoveRedundantQualifierName") // To prevent IDEA replacing FQN imports.
 
-import com.google.common.io.Files.createParentDirs
-import io.spine.internal.dependency.CheckerFramework
-import io.spine.internal.dependency.ErrorProne
-import io.spine.internal.dependency.FindBugs
-import io.spine.internal.dependency.Grpc
-import io.spine.internal.dependency.Guava
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Jackson
 import io.spine.internal.dependency.Protobuf
 import io.spine.internal.dependency.Spine
-import io.spine.internal.dependency.Truth
 import io.spine.internal.gradle.RunBuild
-import io.spine.internal.gradle.VersionWriter
-import io.spine.internal.gradle.checkstyle.CheckStyleConfig
-import io.spine.internal.gradle.excludeProtobufLite
-import io.spine.internal.gradle.forceVersions
-import io.spine.internal.gradle.javac.configureErrorProne
-import io.spine.internal.gradle.javac.configureJavac
-import io.spine.internal.gradle.javadoc.JavadocConfig
-import io.spine.internal.gradle.kotlin.setFreeCompilerArgs
-import io.spine.internal.gradle.publish.IncrementGuard
 import io.spine.internal.gradle.publish.PublishingRepos
 import io.spine.internal.gradle.publish.SpinePublishing
 import io.spine.internal.gradle.publish.spinePublishing
@@ -54,11 +36,7 @@ import io.spine.internal.gradle.report.coverage.JacocoConfig
 import io.spine.internal.gradle.report.license.LicenseReporter
 import io.spine.internal.gradle.report.pom.PomGenerator
 import io.spine.internal.gradle.standardToSpineSdk
-import io.spine.internal.gradle.testing.configureLogging
-import io.spine.internal.gradle.testing.registerTestTasks
 import java.time.Duration
-import java.util.*
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 buildscript {
     standardSpineSdkRepositories()
@@ -80,17 +58,17 @@ buildscript {
 }
 
 plugins {
-    `java-library`
     idea
-    kotlin("jvm")
     errorprone
-    protobuf
+    jacoco
     `gradle-doctor`
-    id(protoData.pluginId) version protoData.version
+    id("project-report")
+    protobuf
+    java
+    id(protoData.pluginId) version protoData.version //apply false
 }
 
-object BuildSettings {
-    const val JAVA_VERSION = 11
+private object BuildSettings {
     const val TIMEOUT_MINUTES = 20L
 }
 
@@ -106,37 +84,19 @@ spinePublishing {
 }
 
 allprojects {
-    apply {
-        plugin("jacoco")
-        plugin("idea")
-        plugin("project-report")
-        apply(from = "$rootDir/version.gradle.kts")
-    }
-
+    apply(from = "$rootDir/version.gradle.kts")
     group = "io.spine.tools"
     version = extra["versionToPublish"]!!
-
     repositories.standardToSpineSdk()
 }
 
 subprojects {
-    applyPlugins()
-    addDependencies()
-    forceConfigurations()
-
-    val javaVersion = JavaLanguageVersion.of(BuildSettings.JAVA_VERSION)
-    configureJava(javaVersion)
-    configureKotlin(javaVersion)
-    setupTests()
-
+    apply(plugin = "module")
+    apply(plugin = "io.spine.protodata")
     setupCodegen()
-
-    val generatedDir = "$projectDir/generated"
-    val generatedResources = "$generatedDir/main/resources"
-    prepareProtocConfigVersionsTask(generatedResources)
-    setupSourceSets(generatedResources)
-
-    configureTaskDependencies()
+    dependencies {
+        protoData(Spine(this).validation.java)
+    }
 }
 
 JacocoConfig.applyTo(project)
@@ -148,14 +108,6 @@ LicenseReporter.mergeAllReports(project)
  * publishing in the root project.
  */
 val publishedModules: Set<String> = extensions.getByType<SpinePublishing>().modules
-
-val testAll by tasks.registering {
-    val testTasks = publishedModules.map { p ->
-        val subProject = project(p)
-        subProject.tasks["test"]
-    }
-    dependsOn(testTasks)
-}
 
 val localPublish by tasks.registering {
     val pubTasks = publishedModules.map { p ->
@@ -177,168 +129,35 @@ val integrationTests by tasks.registering(RunBuild::class) {
 
     /** A timeout for the case of stalled child processes under Windows. */
     timeout.set(Duration.ofMinutes(BuildSettings.TIMEOUT_MINUTES))
-    dependsOn(testAll)
+
+    /** Run integration tests only after all regular tests pass in all modules. */
+    subprojects.forEach {
+        it.tasks.findByName("test")?.let { testTask ->
+            this@registering.dependsOn(testTask)
+        }
+    }
     dependsOn(localPublish)
 }
 
 tasks.register("buildAll") {
-    dependsOn(tasks.build, integrationTests)
+    dependsOn(integrationTests)
 }
 
 val check by tasks.existing {
     dependsOn(integrationTests)
 }
 
-typealias Subproject = Project
+typealias Module = Project
 
-fun Subproject.applyPlugins() {
-    apply {
-        plugin("java-library")
-        plugin("kotlin")
-        plugin("net.ltgt.errorprone")
-        plugin("pmd-settings")
-        plugin(Protobuf.GradlePlugin.id)
-        plugin("io.spine.protodata")
-        plugin("maven-publish")
-        plugin("io.spine.protodata")
-        plugin("detekt-code-analysis")
+/**
+ * We configure ProtoData at the project level (and not in the `module.gradle.kts` script plugin)
+ * because here the Gradle plugin is already applied and its extension is available.
+ */
+fun Module.setupCodegen() {
+
+    protobuf {
+        protoc { artifact = Protobuf.compiler }
     }
-
-    apply<IncrementGuard>()
-    apply<VersionWriter>()
-
-    LicenseReporter.generateReportIn(project)
-    JavadocConfig.applyTo(project)
-    CheckStyleConfig.applyTo(project)
-}
-
-fun Subproject.addDependencies(): Spine {
-    val spine = Spine(project)
-    val validation = spine.validation
-    dependencies {
-        errorprone(ErrorProne.core)
-        protoData(validation.java)
-
-        compileOnlyApi(FindBugs.annotations)
-        compileOnlyApi(CheckerFramework.annotations)
-        ErrorProne.annotations.forEach { compileOnlyApi(it) }
-
-        implementation(Guava.lib)
-
-        testImplementation(Guava.testLib)
-        JUnit.api.forEach { testImplementation(it) }
-        Truth.libs.forEach { testImplementation(it) }
-        testRuntimeOnly(JUnit.runner)
-
-        testImplementation(validation.runtime)
-    }
-    return spine
-}
-
-fun Subproject.forceConfigurations() {
-    val spine = Spine(project)
-    configurations {
-        forceVersions()
-        excludeProtobufLite()
-        all {
-            // Exclude in favor of `spine-validation-java-runtime`.
-            exclude("io.spine", "spine-validate")
-            resolutionStrategy {
-                force(
-                    Protobuf.compiler,
-                    spine.base,
-                    spine.time,
-                    spine.server,
-                    spine.testlib,
-                    spine.toolBase,
-                    spine.pluginBase,
-
-                    // Force the version to avoid the version conflict for
-                    // the `:mc-java:ProtoData` configuration.
-                    spine.validation.runtime,
-                    "io.spine.protodata:protodata-codegen-java:${Spine.protoDataVersion}",
-
-                    JUnit.runner,
-                    "org.hamcrest:hamcrest-core:2.2",
-                    Jackson.core,
-                    Jackson.moduleKotlin,
-                    Jackson.databind,
-                    Jackson.bom,
-                    Jackson.annotations,
-                    Jackson.dataformatYaml,
-
-                    // Transitive dependency.
-                    "io.github.java-diff-utils:java-diff-utils:4.12"
-                )
-            }
-        }
-    }
-}
-
-fun Subproject.configureJava(javaVersion: JavaLanguageVersion) {
-    tasks.withType<JavaCompile>().configureEach {
-        sourceCompatibility = javaVersion.toString()
-        targetCompatibility = javaVersion.toString()
-        configureJavac()
-        configureErrorProne()
-    }
-}
-
-fun Subproject.configureKotlin(javaVersion: JavaLanguageVersion) {
-    kotlin {
-        explicitApi()
-    }
-    tasks.withType<KotlinCompile>().configureEach {
-        kotlinOptions.jvmTarget = javaVersion.toString()
-        setFreeCompilerArgs()
-    }
-}
-
-fun Subproject.setupTests() {
-    tasks {
-        registerTestTasks()
-        withType<Test> {
-            useJUnitPlatform()
-            configureLogging()
-        }
-    }
-}
-
-fun Subproject.prepareProtocConfigVersionsTask(generatedResources: String) {
-    val prepareProtocConfigVersions by tasks.registering {
-        description = "Prepares the versions.properties file."
-
-        val propertiesFile = file("$generatedResources/versions.properties")
-        outputs.file(propertiesFile)
-
-        val versions = Properties().apply {
-            setProperty("baseVersion", Spine.DefaultVersion.base)
-            setProperty("protobufVersion", Protobuf.version)
-            setProperty("gRPCVersion", Grpc.version)
-        }
-
-        @Suppress("UNCHECKED_CAST")
-        inputs.properties(HashMap(versions) as MutableMap<String, *>)
-
-        doLast {
-            createParentDirs(propertiesFile)
-            propertiesFile.createNewFile()
-            propertiesFile.outputStream().use {
-                versions.store(it,
-                    "Versions of dependencies of the Spine Model Compiler for Java plugin and" +
-                            " the Spine Protoc plugin.")
-            }
-        }
-    }
-
-    tasks.processResources {
-        dependsOn(prepareProtocConfigVersions)
-    }
-}
-
-fun Subproject.setupCodegen() {
-
-    protobuf { protoc { artifact = Protobuf.compiler } }
 
     protoData {
         renderers(
@@ -352,11 +171,5 @@ fun Subproject.setupCodegen() {
         plugins(
             "io.spine.validation.ValidationPlugin"
         )
-    }
-}
-
-fun Subproject.setupSourceSets(generatedResources: String) {
-    sourceSets.main {
-        resources.srcDir(generatedResources)
     }
 }
